test: reorganisation of unit tests: conn_req
