+++ tests/unit/common/test-common.h
@@ -17,9 +17,9 @@
 #define MOCK_PEER		(struct rpma_peer *)0xFEEF
 #define MOCK_INFO		(struct rpma_info *)0xE6B2
 #define MOCK_CONN		(struct rpma_conn *)0xC004
-#define MOCK_PRIVATE_DATA	((void *)"Random data")
+#define MOCK_PRIVATE_DATA	((void *)Private_data)
 #define MOCK_PDATA_LEN		(strlen(MOCK_PRIVATE_DATA) + 1)
-#define MOCK_PRIVATE_DATA_2	((void *)"Another random data")
+#define MOCK_PRIVATE_DATA_2	((void *)Private_data_2)
 #define MOCK_PDATA_LEN_2	(strlen(MOCK_PRIVATE_DATA_2) + 1)
 #define MOCK_ERRNO		123456
 
@@ -28,5 +28,8 @@
 #define MOCK_PASSTHROUGH	0
 #define MOCK_VALIDATE		1
 #define MOCK_ERR_PENDING	(-1)
+
+extern const char Private_data[];
+extern const char Private_data_2[];
 
 #endif /* TEST_COMMON_H */
